Split launcher crate on 'database', 'launcher', and 'tracing' features
