+++ libs/blockscout-service-launcher/src/launcher/mod.rs
@@ -1,0 +1,8 @@
+mod launch;
+mod metrics;
+mod router;
+mod settings;
+
+pub use launch::{launch, LaunchSettings};
+pub use router::HttpRouter;
+pub use settings::*;
