+++ libs/blockscout-service-launcher/src/tracing/mod.rs
@@ -1,0 +1,5 @@
+mod init;
+mod settings;
+
+pub use init::*;
+pub use settings::*;
